refactor: addMember와 관련된 책임을 CellManager에 옮기고 관련된 테스트를 수정
