+++ src/test/java/com/windstorm/management/implement/cell/CellAppenderTest.java
@@ -30,12 +30,6 @@ class CellAppenderTest {
 	private CellRepository cellRepository;
 
 	@Mock
-	private MemberReader memberReader;
-
-	@Mock
-	private CellReader cellReader;
-
-	@Mock
 	private UnitReader unitReader;
 
 	@InjectMocks
@@ -84,88 +78,5 @@ class CellAppenderTest {
 		assertThatThrownBy(() -> cellAppender.append(request))
 			.isInstanceOf(RuntimeException.class)
 			.hasMessage(unitName + "에 해당하는 진이 없습니다.");
-	}
-
-	@Test
-	@DisplayName("셀에 셀원을 추가할 수 있다.")
-	void addMemberSuccess() {
-		// given
-		CellAddMember request = CellAddMember.builder()
-			.uniqueId("1")
-			.cellName("홍길동셀")
-			.build();
-
-		Member member = createMember();
-		Cell cell = createCell();
-
-		when(memberReader.read(any(String.class))).thenReturn(member);
-		when(cellReader.read(any(String.class))).thenReturn(cell);
-
-		// when
-		cellAppender.addMember(request);
-
-		// then
-		assertThat(cell.getMembers())
-			.hasSize(1)
-			.extracting("uniqueId", "name")
-			.containsExactlyInAnyOrder(
-				tuple("1", "김철수")
-			);
-	}
-
-	@Test
-	@DisplayName("존재하지 않는 유저는 셀에 추가할 수 없다.")
-	void addMemberFail1() {
-		// given
-		String uniqueId = "100";
-
-		CellAddMember request = CellAddMember.builder()
-			.uniqueId("1")
-			.cellName("홍길동셀")
-			.build();
-
-		when(memberReader.read(any(String.class))).thenThrow(new RuntimeException("교적번호: " + uniqueId + "에 해당하는 데이터가 존재하지 않습니다."));
-
-		// when, then
-		assertThatThrownBy(() -> cellAppender.addMember(request))
-			.isInstanceOf(RuntimeException.class)
-			.hasMessage("교적번호: " + uniqueId + "에 해당하는 데이터가 존재하지 않습니다.");
-	}
-
-	@Test
-	@DisplayName("존재하지 않는 셀에 유저를 추가할 수 없다.")
-	void addMemberFail2() {
-		// given
-		String cellName = "김짱구셀";
-
-		CellAddMember request = CellAddMember.builder()
-			.uniqueId("1")
-			.cellName("홍길동셀")
-			.build();
-
-		when(cellReader.read(any(String.class))).thenThrow(new RuntimeException(cellName + "에 해당하는 셀이 존재하지 않습니다."));
-
-		// when, then
-		assertThatThrownBy(() -> cellAppender.addMember(request))
-			.isInstanceOf(RuntimeException.class)
-			.hasMessage(cellName + "에 해당하는 셀이 존재하지 않습니다.");
-	}
-
-	private Member createMember() {
-		return Member.create(
-			"1",
-			"김철수",
-			"saeeden1234!!",
-			LocalDate.of(2000, 11, 20),
-			Division.DANIEL,
-			Gender.MALE,
-			LeaderRole.CAPTAIN,
-			"010-1234-5678",
-			"경기도 성남시"
-		);
-	}
-
-	private Cell createCell() {
-		return Cell.create("홍길동셀", Division.DANIEL);
 	}
 }
